include change log id in data and current tables
This commit modifies the storage:

1) to include reference to the change log id in the current table by extending
the values in this section of the db to be version+SEP+change.

2) and to include reference to the change log id in the data table, by
including it in the key. The key format is modified to
PREFIX_DATA+subset+SEP+key+SEP+version+SEP+change.

put, get and friends must all now utilize the change information when pulling a
specific object. Appropriate modifications have been implemented.

Although functionality based on this change to the storage model has not been
tested, the existing tests in dat (20f3ed5c3) pass without error (616/616).